+++ runner/src/main/java/com/walmartlabs/concord/runner/engine/VariablesTask.java
@@ -1,10 +1,11 @@
 package com.walmartlabs.concord.runner.engine;
 
+import com.walmartlabs.concord.common.InjectVariable;
 import com.walmartlabs.concord.common.Task;
 import io.takari.bpm.api.ExecutionContext;
 
 import javax.inject.Named;
-import java.util.Map;
+import java.util.*;
 
 @Named("vars")
 public class VariablesTask implements Task {
@@ -24,5 +25,24 @@ public class VariablesTask implements Task {
 
     public void set(ExecutionContext ctx, Map<String, Object> vars) {
         vars.forEach(ctx::setVariable);
+    }
+
+    public Object eval(@InjectVariable("execution") ExecutionContext ctx, Object v) {
+        return ctx.interpolate(v);
+    }
+
+    public List<Object> concat(Collection<Object> a, Collection<Object> b) {
+        if (a == null) {
+            a = Collections.emptyList();
+        }
+
+        if (b == null) {
+            b = Collections.emptyList();
+        }
+
+        List<Object> l = new ArrayList<>(a.size() + b.size());
+        l.addAll(a);
+        l.addAll(b);
+        return l;
     }
 }